Rename LandingPageHeroSection and HeroSection
