+++ services/cms/src/blocks/HeroSection/index.tsx
@@ -3,22 +3,11 @@ import { BlockConfiguration } from "@wordpress/blocks"
 import HeroSectionEditor from "./HeroSectionEditor"
 import HeroSectionSave from "./HeroSectionSave"
 
-export interface HeroSectionAttributes {
-  heroHeading: string
-}
-
-const HeroSectionConfiguration: BlockConfiguration<HeroSectionAttributes> = {
+const HeroSectionConfiguration: BlockConfiguration<Record<string, never>> = {
   title: "Hero Section",
-  description:
-    "A hero section is a full screen section typically consisting of a background image, or animations, with text and sometimes a call to action.",
+  description: "A hero section for chapter front page with a heading and subheading.",
   category: "design",
-  attributes: {
-    heroHeading: {
-      type: "string",
-      source: "html",
-      selector: "h1",
-    },
-  },
+  attributes: {},
   edit: HeroSectionEditor,
   save: HeroSectionSave,
 }
