+++ services/cms/src/blocks/LandingPageHeroSection/index.tsx
@@ -1,0 +1,26 @@
+import { BlockConfiguration } from "@wordpress/blocks"
+
+import LandingPageHeroSectionEditor from "./LandingPageHeroSectionEditor"
+import LandingPageHeroSectionSave from "./LandingPageHeroSectionSave"
+
+export interface LandingPageHeroSectionAttributes {
+  title: string
+}
+
+const LandingPageHeroSectionConfiguration: BlockConfiguration<LandingPageHeroSectionAttributes> = {
+  title: "Landing Page Hero Section",
+  description:
+    "Landing page hero section is a full screen section typically consisting of a background image, or animations, with text and sometimes a call to action.",
+  category: "design",
+  attributes: {
+    title: {
+      type: "string",
+      source: "html",
+      selector: "h1",
+    },
+  },
+  edit: LandingPageHeroSectionEditor,
+  save: LandingPageHeroSectionSave,
+}
+
+export default LandingPageHeroSectionConfiguration
