+++ services/cms/src/blocks/index.tsx
@@ -1,6 +1,5 @@
 import { BlockConfiguration } from "@wordpress/blocks"
 
-import ChapterHeroSection from "./ChapterHeroSection"
 import ChapterProgress from "./ChapterProgress"
 import CourseChapterGrid from "./CourseChapterGrid"
 import CourseObjectiveSection from "./CourseObjectiveSection"
@@ -9,6 +8,7 @@ import Exercise from "./Exercise"
 import ExerciseInChapter from "./ExerciseInChapter"
 import ExerciseTask from "./ExerciseTask"
 import HeroSection from "./HeroSection"
+import LandingPageHeroSection from "./LandingPageHeroSection"
 import Latex from "./Latex"
 import PagesInChapter from "./PagesInChapter"
 
@@ -26,7 +26,6 @@ export const blockTypeMapForPages: Array<[string, BlockConfiguration<any>]> = [
   ["moocfi/chapter-progress", ChapterProgress],
   ["moocfi/latex", Latex],
   ["moocfi/hero-section", HeroSection],
-  ["moocfi/chapter-hero-section", ChapterHeroSection],
 ]
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -34,6 +33,6 @@ export const blockTypeMapForTopLevelPages: Array<[string, BlockConfiguration<any
   ["moocfi/course-chapter-grid", CourseChapterGrid],
   ["moocfi/course-progress", CourseProgress],
   ["moocfi/latex", Latex],
-  ["moocfi/hero-section", HeroSection],
+  ["moocfi/landing-page-hero-section", LandingPageHeroSection],
   ["moocfi/course-objective-section", CourseObjectiveSection],
 ]
