+++ services/headless-lms/src/models/chapters.rs
@@ -303,7 +303,7 @@ RETURNING *;
     .await?;
 
     let chapter_frontpage_content = serde_json::to_value(vec![
-        GutenbergBlock::empty_block_from_name("moocfi/chapter-hero-section".to_string()),
+        GutenbergBlock::empty_block_from_name("moocfi/hero-section".to_string()),
         GutenbergBlock::empty_block_from_name("moocfi/pages-in-chapter".to_string()),
         GutenbergBlock::empty_block_from_name("moocfi/chapter-progress".to_string()),
         GutenbergBlock::empty_block_from_name("moocfi/exercises-in-chapter".to_string()),
