+++ services/headless-lms/src/models/courses.rs
@@ -141,7 +141,7 @@ pub async fn insert_course(
 
     // Create front page for course
     let course_front_page_content = serde_json::to_value(vec![
-        GutenbergBlock::empty_block_from_name("moocfi/hero-section".to_string()),
+        GutenbergBlock::empty_block_from_name("moocfi/landing-page-hero-section".to_string()),
         GutenbergBlock::empty_block_from_name("moocfi/course-objective-section".to_string()),
         GutenbergBlock::empty_block_from_name("moocfi/course-chapter-grid".to_string()),
         GutenbergBlock::empty_block_from_name("moocfi/course-progress".to_string()),
